% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - As date.R
\name{f_as_date}
\alias{f_as_date}
\title{Convert string to date.}
\usage{
f_as_date(v.input, c.format = "\%d/\%m/\%Y", c.origin = "1899-12-30")
}
\arguments{
\item{v.input}{Vector with strings to be converted.}

\item{c.format}{Format of date in the strings (default: "\%d/\%m/\%Y").}

\item{c.origin}{Starting date (default: "1899-12-30")}
}
\value{
Vector with dates.
}
\description{
Converts a string to a date.
}
\details{
-
}
\examples{
v.output <- f_as_date(
           v.input  = c("16/03/2022", "16/03/2022", "16/03/2022"),
           c.format = "\%d/\%m/\%Y",
           c.origin = "1899-12-30"
)
}
\author{
Pieter Overdevest
}
