% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - As integer.R
\name{f_as_integer}
\alias{f_as_integer}
\title{Convert string columns to integer columns.}
\usage{
f_as_integer(df.input, v.col, c.id)
}
\arguments{
\item{df.input}{Data frame with columns you want to convert from char to integer.}

\item{v.col}{Names of columns you want to convert from char to integer.}

\item{c.id}{Column name with unique identifier for each row. This is used to report on parsing errors.}
}
\value{
Data frame with updated columns.
}
\description{
Converts data frame columns of type character to integers.
}
\details{
-
}
\examples{
df.output <- f_as_integer(

           df.input  = tibble(id = c("A", "B"), id.product = c("1", "2"), id.variant = c("32", "45")),
           v.col     = c("id.product", "id.variant"),
           c.id      = "id"
)
}
\author{
Pieter Overdevest
}
