% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Calculate Std Error.R
\name{f_calculate_std_error}
\alias{f_calculate_std_error}
\title{Calculate standard error}
\usage{
f_calculate_std_error(df.input, c.group.by = NULL, c.model, c.data)
}
\arguments{
\item{df.input}{-----}

\item{c.group.by}{----- (default: NULL).}

\item{c.model}{-----}

\item{c.data}{-----}
}
\value{
Vector with standard errors.
}
\description{
Calculates standard error.
}
\details{
-
}
\examples{
v.output <- f_calculate_std_error(

     df.input   = df.sc,
     c.group.by = NULL,
     c.model    = "solids.median",
     c.data     = "solids"

)
}
\author{
Pieter Overdevest
}
