% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Check col not empty.R
\name{f_check_col_not_empty}
\alias{f_check_col_not_empty}
\title{Check whether column is empty}
\usage{
f_check_col_not_empty(df.input, c.col)
}
\arguments{
\item{df.input}{-----}

\item{c.col}{-----}
}
\value{
Nothing.
}
\description{
Checks whether column is empty.
}
\details{
-
}
\examples{
f_check_col_not_empty(

       df.input,
       c.col
)
}
\author{
Pieter Overdevest
}
