% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Check col unique.R
\name{f_check_col_unique}
\alias{f_check_col_unique}
\title{Check whether column is unique}
\usage{
f_check_col_unique(df.input, c.col)
}
\arguments{
\item{df.input}{-----}

\item{c.col}{-----}
}
\value{
Nothing.
}
\description{
Checks whether column is unique.
}
\details{
-
}
\examples{
f_check_col_unique(

       df.input,
       c.col
)
}
\author{
Pieter Overdevest
}
