% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Check cols not zero.R
\name{f_check_cols_not_zero}
\alias{f_check_cols_not_zero}
\title{Check whether columns are not zero}
\usage{
f_check_cols_not_zero(
  df.input,
  v.col.include = NULL,
  v.col.exclude = NULL,
  c.id
)
}
\arguments{
\item{df.input}{-----}

\item{v.col.include}{----- (default: NULL)}

\item{v.col.exclude}{----- (default: NULL)}

\item{c.id}{-----}
}
\value{
Nothing.
}
\description{
Checks whether columns are not zero.
}
\details{
-
}
\examples{
f_check_cols_not_zero(

    df.input,
    v.col.include = NULL,
    v.col.exclude = NULL,
    c.id
)
}
\author{
Pieter Overdevest
}
