% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Check cols present.R
\name{f_check_cols_present}
\alias{f_check_cols_present}
\title{Check whether columns are present}
\usage{
f_check_cols_present(df.input, v.col)
}
\arguments{
\item{df.input}{-----}

\item{c.id}{-----}
}
\value{
Nothing.
}
\description{
Checks whether columns are present.
}
\details{
-
}
\examples{
f_check_cols_present(

    df.input,
    v.col
)
}
\author{
Pieter Overdevest
}
