% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Check cols unique.R
\name{f_check_cols_unique}
\alias{f_check_cols_unique}
\title{Check whether columns are unique}
\usage{
f_check_cols_unique(df.input, v.col.include = NULL, v.col.exclude = NULL, c.id)
}
\arguments{
\item{df.input}{-----}

\item{v.col.include}{----- (default: NULL)}

\item{v.col.exclude}{----- (default: NULL)}

\item{c.id}{-----}
}
\value{
Nothing.
}
\description{
Checks whether columns are unique.
}
\details{
-
}
\examples{
f_check_cols_unique(

    df.input,
    v.col.include = NULL,
    v.col.exclude = NULL,
    c.id
)
}
\author{
Pieter Overdevest
}
