% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Compare two vectors.R
\name{f_compare_two_vectors}
\alias{f_compare_two_vectors}
\title{Compare two vectors}
\usage{
f_compare_two_vectors(v1, v2)
}
\arguments{
\item{v1, v2}{Vectors with elements.}
}
\value{
Nothing.
}
\description{
Compares two vectors.
}
\details{
-
}
\examples{
f_compare_two_vectors(

    v1 = c(1,2,1),
    v2 = c(1,1,2)
)
}
\author{
Pieter Overdevest
}
