% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Define Col Type.R
\name{f_define_col_type}
\alias{f_define_col_type}
\title{Define column types}
\usage{
f_define_col_type(v.names, v.double, v.integer, v.date.time)
}
\arguments{
\item{v.names}{-----}

\item{v.double}{-----}

\item{v.integer}{-----}

\item{v.date.time}{-----}
}
\value{
List of column types.
}
\description{
Defines column type.
}
\details{
-
}
\examples{
l.col.type <- f_define_col_type(

    df.input,
    v.col.include = NULL,
    v.col.exclude = NULL,
    c.id
)
}
\author{
Pieter Overdevest
}
