% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Distance.R
\name{f_distance}
\alias{f_distance}
\title{Calculate distances}
\usage{
f_distance(x1, y1, x2, y2)
}
\arguments{
\item{x1, y1}{Coordinates of first point.}

\item{x2, y2}{Coordinates of second point.}
}
\value{
Vector with distances.
}
\description{
Calculates distances.
}
\details{
-
}
\examples{
v.distance <- f_distance(

    x1 = c(1 ,1),
    y1 = c(1, 2),
    x2 = c(3, 3),
    y2 = c(4, 5)
)
}
\author{
Pieter Overdevest
}
