% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Geom Point.R
\name{f_geom_point}
\alias{f_geom_point}
\title{Plot geom point}
\usage{
f_geom_point(
  df.plot,
  c.x,
  c.y,
  c.facet = NULL,
  n.alpha = 0.2,
  b.log.x = FALSE,
  b.log.y = FALSE,
  v.abline = NULL,
  c.labs.x = NULL,
  c.labs.y = NULL,
  b.grid.major.x = FALSE,
  b.grid.major.y = FALSE,
  b.coord.fixed = FALSE,
  v.scale.x.continuous = NULL,
  v.scale.y.continuous = NULL,
  c.labs.title = waiver(),
  c.labs.subtitle = waiver(),
  c.labs.caption = waiver()
)
}
\arguments{
\item{df.plot}{Data frame with data to plot.}

\item{c.x}{Column name in said data frame to plot on the x-axis.}

\item{c.y}{Column name in said data frame to plot on the y-axis.}

\item{n.alpha}{Transparancy of points (default: 0.2)}

\item{b.log.x}{Should x-scale be log10? (default: FALSE)}

\item{b.log.y}{Should y-scale be log10? (default: FALSE)}

\item{v.abline}{Slope and intercept in a vector (default: NULL)}

\item{c.labs.x}{Label for x-axis (default: NULL)}

\item{c.labs.y}{Label for y-axis (default: NULL)}

\item{b.grid.major.x}{Should grid added on x-axis (default: FALSE)}

\item{b.grid.major.y}{Should grid added on y-axis (default: FALSE)}

\item{b.coord.fixed}{Should x-axis and y-axis have fixed ratio of 1 (default: FALSE)}

\item{v.scale.x.continuous}{Breaks along x-axis (default: NULL)}

\item{v.scale.y.continuous}{Breaks along y-axis (default: NULL)}

\item{c.labs.title}{Main title (default: waiver())}

\item{c.labs.subtitle}{Subtitle (default: waiver())}

\item{c.labs.caption}{Caption under the plot (default: waiver())}
}
\value{
A ggplot object that can be printed.
}
\description{
Plot geom point
}
\details{
-
}
\examples{
f_geom_point(

       df.plot         df.plot,
       c.x             = 'columnx',
       c.y             = 'columny',
       n.alpha         = 0.2,
       b.log.x         = FALSE,
       b.log.y         = FALSE,
       v.abline        = NULL,
       c.labs.x        = NULL,
       c.labs.y        = NULL,
       c.labs.title    = waiver(),
       c.labs.subtitle = waiver(),
       c.labs.caption  = waiver()
)
}
\author{
Pieter Overdevest
}
