% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Get filenames in folder.R
\name{f_get_filenames_in_folder}
\alias{f_get_filenames_in_folder}
\title{Get names and information on files in folder}
\usage{
f_get_filenames_in_folder(
  c.path,
  b.recursive = FALSE,
  c.file.type = NULL,
  b.return.md5 = FALSE
)
}
\arguments{
\item{c.path}{Full path to folder of interest.}

\item{b.recursive}{Should we look inside subfolders (default: FALSE).}

\item{c.file.type}{What file type to retrieve? (default: NULL).}

\item{b.return.md5}{Should MD5 be returned? (default: FALSE).}
}
\value{
Data frame with information on the files observed in concerned folder.
}
\description{
Get names and information on files in folder.
}
\details{
-
}
\examples{
df.output <- f_get_filenames_in_folder(

    c.path,
    b.recursive,
    c.file.type,
    b.return.md5
)
}
\author{
Pieter Overdevest
}
