% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Get index from list.R
\name{f_get_index_from_list}
\alias{f_get_index_from_list}
\title{Get index of searched items in list}
\usage{
f_get_index_from_list(l.input, v.element, v.search)
}
\arguments{
\item{l.input}{-----}

\item{v.element}{-----}

\item{v.search}{-----}
}
\value{
Vector with indices of elements that match searched information.
}
\description{
Get index of searched items in list.
}
\details{
-
}
\examples{
v.output <- f_get_index_from_list(

    l.input,
    v.element,
    v.search
)
}
\author{
Pieter Overdevest
}
