% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Get latest file.R
\name{f_get_latest_file}
\alias{f_get_latest_file}
\title{Get latest file}
\usage{
f_get_latest_file(
  c.file.string,
  c.file.type,
  c.path,
  b.exact.match = FALSE,
  c.file.string.exclude = NULL,
  c.sheet.name = NULL,
  c.show.report = "all"
)
}
\arguments{
\item{c.file.string}{String to search for in the filename.}

\item{c.file.type}{File type.}

\item{c.path}{Path where the files should be searched for.}

\item{b.exact.match}{Should the string match the files exactly? (default: FALSE)}

\item{c.file.string.exclude}{String in files to exclude (default: NULL)}

\item{c.sheet.name}{Sheetname, in case of Excel file. The first sheet will be chosen in case the value is NULL (default: NULL)}

\item{c.show.report}{What feedback should we show? Options: 'none', 'minimal', 'all' (default: 'all')}
}
\value{
Full path to latest file in concerned folder.
}
\description{
Get latest file with given c.file.string and of type c.file.type.
}
\details{
-
}
\examples{
c.path.file <- f_get_latest_file(

       c.file.string         = "allocation",
       c.file.type           = "xls",
       c.path                = path.data,
       b.exact.match         = FALSE,
       c.file.string.exclude = NULL,
       c.sheet.name          = NULL,
       c.show.report         = "all"
)
}
\author{
Pieter Overdevest
}
