% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - GS col number.R
\name{f_gs_col_number}
\alias{f_gs_col_number}
\title{Get number of columns in Google Sheet}
\usage{
f_gs_col_number(c.gs.code, c.sheet = "Sheet1")
}
\arguments{
\item{c.gs.code}{Google Sheet code.}

\item{c.sheet}{Sheet name (default: "Sheet1")}
}
\value{
Number of columns.
}
\description{
Get number of columns in Google Sheet.
}
\details{
-
}
\examples{
n.col <- f_gs_col_number(

      c.gs.code = "1Gc7Z1fxLgz60xg5wGpROMyfdwubk9mqeFaMzXh6WIjU"
      c.sheet   = "Delta"
)
}
\author{
Pieter Overdevest
}
