% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - GS url.R
\name{f_gs_url}
\alias{f_gs_url}
\title{Get URL to Google Sheet}
\usage{
f_gs_url(c.gs.code, c.sheet = "Sheet1")
}
\arguments{
\item{c.gs.code}{Google Sheet code.}

\item{c.sheet}{Sheet name (default: "Sheet1")}
}
\value{
URL to Google Sheet.
}
\description{
Get URL to Google Sheet.
}
\details{
-
}
\examples{
c.url <- f_gs_url(

      c.gs.code = "1Gc7Z1fxLgz60xg5wGpROMyfdwubk9mqeFaMzXh6WIjU"
      c.sheet   = "Delta"
)
}
\author{
Pieter Overdevest
}
