% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Install package from CRAN.R
\name{f_install_package_from_cran}
\alias{f_install_package_from_cran}
\title{Install latest packages from CRAN}
\usage{
f_install_package_from_cran(v.package)
}
\arguments{
\item{v.package}{Vector with package names to install}
}
\value{
Nothing.
}
\description{
Install latest packages from CRAN, and 'library' them.
}
\details{
-
}
\examples{
f_install_package_from_cran(

    v.package = c("tmaptools")
)
}
\author{
Pieter Overdevest
}
