% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Install package from GitHub.R
\name{f_install_package_from_github}
\alias{f_install_package_from_github}
\title{Install last version of package on GitHub}
\usage{
f_install_package_from_github(c.package, c.auth.token = NULL)
}
\arguments{
\item{c.package}{Package name.}

\item{c.auth.token}{Authentication token to access private repo. (default: NULL)}
}
\value{
Nothing.
}
\description{
Install last version of package on GitHub.
}
\details{
Note - Never insert a token in the script. GitHub recognizes these tokens when committing and
revokes the concerned tokens. Great for security, bit unhandy in use!
}
\examples{
f_install_package_from_github(

    c.package    = "generaltoolbox"
    c.auth.token = NULL
)
}
\author{
Pieter Overdevest
}
