% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Is same.R
\name{f_is_same}
\alias{f_is_same}
\title{Are values in two vectors/columns the same.}
\usage{
f_is_same(x1, x2)
}
\arguments{
\item{x1}{Column or vector 1.}

\item{x2}{Column or vector 2.}
}
\value{
Column containing TRUE or FALSE.
}
\description{
Check per index/row whether values in two vectors/columns are the same.
}
\details{
-
}
\examples{
f_is_same(
       x1 = c(1,2,3,4,5,NA),
       x2 = c(1,2,3,4,NA,NA)
)

}
\author{
Pieter Overdevest
}
