% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Is unique.R
\name{f_is_unique}
\alias{f_is_unique}
\title{Check whether column is unique}
\usage{
f_is_unique(v.vector, v.excluding = NA)
}
\arguments{
\item{v.vector}{Vector with elements.}

\item{v.excluding}{Vector with elements to exclude, e.g., 0, NA or NULL (default: NA)}
}
\value{
Boolean whether the vector contains unique values.
}
\description{
Are all elements in vector unique?
}
\details{
-
}
\examples{
b.unique <- f_is_unique(

    v.vector    = c(1, 2, 3, NA),    ,
    v.excluding = NA
)
}
\author{
Pieter Overdevest
}
