% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Kable.R
\name{f_kable}
\alias{f_kable}
\title{Print Kable in R-Markdown}
\usage{
f_kable(
  df.input,
  c.caption = "Add nice caption through 'c.caption'",
  c.position = "center",
  v.align = NULL,
  v.width.cm = 2,
  n.angle = NULL,
  n.font.size = 8,
  c.latex_options = "basic",
  v.grey.col = NULL,
  b.grey.col = TRUE,
  n.top = "all"
)
}
\arguments{
\item{df.input}{Data frame to put in kable table.}

\item{c.caption}{Add nice caption.}

\item{c.position}{Options: "center", "left" (default: "center")}

\item{v.align}{Vector containing "l", "c", "r" indicating how to align each column (default: NULL)}

\item{v.width.cm}{Width of column (default: 2)}

\item{n.angle}{Angle of text in header (default: NULL).}

\item{n.font.size}{Font size (default: 8).}

\item{c.latex_options}{String of latex options (default: "basic").}

\item{v.grey.col}{Column numbers to give grey background (default: NULL).}

\item{b.grey.col}{Add grey column to columns with total in header? (default: TRUE).}

\item{n.top}{Number of rows to print (default: "all").}
}
\value{
Kable output.
}
\description{
Prints Kable in R-Markdown (PDF only, not HTML).
}
\details{
-
}
\examples{
kable.output <- f_kable(

       df.input            = mtcars,
       c.caption           = "Add nice caption",
       c.position          = "center",
       v.align             = NULL,
       v.width.cm          = 2,
       n.angle             = NULL,
       n.font.size         = 8,
       c.latex_options     = "basic",
       v.grey.col          = NULL,
       b.grey.col          = TRUE,
       n.top               = "all"
)
}
\author{
Pieter Overdevest
}
