% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - lapply.R
\name{f_lapply}
\alias{f_lapply}
\title{Apply function to list of items}
\usage{
f_lapply(l.input, f.input, mc.cores = NULL)
}
\arguments{
\item{l.input}{List of items.}

\item{f.input}{Function to apply to each item.}

\item{mc.cores}{Number of cores (default: NULL)}
}
\value{
List of items that have been processed with said function.
}
\description{
Applies function to list of items.
}
\details{
-
}
\examples{
l.output <- f_lapply(

     l.input  = list(1, 2, 3),
     f.input  = function(x) {x+2},
     mc.cores = NULL
)
}
\author{
Pieter Overdevest
}
