% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Names.R
\name{f_names}
\alias{f_names}
\title{Give names of data frame.}
\usage{
f_names(df.input, b.sort.name = TRUE, b.sort.class = TRUE)
}
\arguments{
\item{df.input}{Data frame to get names of.}

\item{b.sort.name}{Should names be sorted? (default: TRUE).}

\item{b.sort.class}{Should types be sorted? (default: TRUE).}
}
\value{
Nothing.
}
\description{
Gives names of data frame.
}
\details{
-
}
\examples{
f_names(

       df.input     = mtcars,
       b.sort.name  = TRUE,
       b.sort.class = TRUE
)
}
\author{
Pieter Overdevest
}
