% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Object size.R
\name{f_object_size}
\alias{f_object_size}
\title{Calculate object size}
\usage{
f_object_size(df.input)
}
\arguments{
\item{df.input}{Object to get size of.}
}
\value{
Size of object in KB, MB, GB, depending of size.
}
\description{
Calculate object size.
}
\details{
-
}
\examples{
f_object_size(

    df.input = mtcars
)
}
\author{
Pieter Overdevest
}
