% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - OCR table data from PDF.R
\name{f_ocr_table_data_from_pdf}
\alias{f_ocr_table_data_from_pdf}
\title{OCR table data from PDF}
\usage{
f_ocr_table_data_from_pdf(c.file.string, c.path.source, c.path.destination)
}
\arguments{
\item{c.file.string}{Filename (or part thereof) of PDF to be OCR-ed.}

\item{c.path.source}{Path where PDF file can be found.}

\item{c.path.destination}{Path were resulting XLS should be saved.}
}
\value{
List of OCR-ed text
}
\description{
OCR table data from PDF.
}
\details{
-
}
\examples{
l.text <- f_ocr_table_data_from_pdf(

       c.file.string,
       c.path.source,
       c.path.destination
)
}
\author{
Pieter Overdevest
}
