% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Paste.R
\name{f_paste}
\alias{f_paste}
\title{Concatenate items in vector}
\usage{
f_paste(
  v.string,
  c.collapse = ",",
  c.and = "and",
  b.capitalize = FALSE,
  b.quotation = FALSE,
  b.sort = TRUE,
  b.unique = FALSE,
  n.top = NULL
)
}
\arguments{
\item{v.string}{Vector of items to concatenate.}

\item{c.collapse}{(default: ",").}

\item{c.and}{(default: "and").}

\item{b.capitalize}{(default: FALSE).}

\item{b.quotation}{(default: FALSE).}

\item{b.sort}{(default: TRUE).}

\item{b.unique}{(default: FALSE).}

\item{n.top}{(default: NULL).}
}
\value{
Concatenated string.
}
\description{
Concatenate items in vector.
}
\details{
-
}
\examples{
c.string <- f_paste(

       v.string     = c("A", "B", "C"),
       c.collapse   = ",",
       c.and        = "and",
       b.capitalize = FALSE,
       b.quotation  = FALSE,
       b.sort       = TRUE,
       b.unique     = FALSE,
       n.top        = NULL
)
}
\author{
Pieter Overdevest
}
