% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Paste regex.R
\name{f_paste_regex}
\alias{f_paste_regex}
\title{Concatenate regex items in vector separated by "|"}
\usage{
f_paste_regex(
  v.string,
  c.pre = "^",
  c.collapse = "|",
  c.post = "$",
  b.sort = TRUE
)
}
\arguments{
\item{v.string}{Vector with items.}

\item{c.pre}{String to pre-pend (default: "^").}

\item{c.collapse}{String to collapse items (default: "|").}

\item{c.post}{String to append (default: "$").}

\item{b.sort}{Should items be sorted? (default: TRUE).}
}
\value{
.
}
\description{
Concatenate regex items in vector separated by "|".
}
\details{
-
}
\examples{
c.string <- f_paste_regex(

       v.string     = c("pieter", "bart", "theo", "kees", "aad"),
       c.pre        = "^",
       c.collapse   = "|",
       c.post       = "$",
       b.sort       = TRUE
)
}
\author{
Pieter Overdevest
}
