% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Plot leaflet.R
\name{f_plot_leaflet}
\alias{f_plot_leaflet}
\title{Plot leaflet}
\usage{
f_plot_leaflet(
  c.leaflet.title = "",
  c.legend.title = "",
  c.layer = NULL,
  v.coord.point.midpoint = NULL,
  n.zoom = 13,
  b.save.leaflet = FALSE,
  b.add.date = TRUE,
  b.add.time = TRUE,
  b.add.streetsmart = TRUE,
  df.point = NULL,
  v.coord.point = c("bord.lon", "bord.lat"),
  c.id.point = "bord.id",
  b.show.point.label = TRUE,
  v.info.tag.point.label = NULL,
  v.info.veld.point.label = NULL,
  b.show.point.popup = TRUE,
  v.info.tag.point.popup = NULL,
  v.info.veld.point.popup = NULL,
  df.fill.point = NULL,
  c.fill.factor.point = NULL,
  c.fill.numeric.point = NULL,
  df.weight.point = NULL,
  c.weight.point = NULL,
  df.stroke.color.point = NULL,
  c.stroke.factor.point = NULL,
  c.stroke.numeric.point = NULL,
  df.stroke.weight.point = NULL,
  c.stroke.weight.point = NULL,
  n.opacity.fill = 0.9,
  n.opacity.stroke = 0.9,
  df.line = NULL,
  v.coord.line = c("line.lon", "line.lat"),
  c.id.line = "line.id",
  b.show.line.label = FALSE,
  v.info.tag.line.label = NULL,
  v.info.veld.line.label = NULL,
  b.show.line.popup = FALSE,
  v.info.tag.line.popup = NULL,
  v.info.veld.line.popup = NULL,
  df.color.line = NULL,
  c.color.line = NULL,
  df.weight.line = NULL,
  c.weight.line = NULL,
  n.opacity.line = 1,
  df.polygon = NULL,
  v.coord.polygon = c("polygon.lon", "polygon.lat"),
  c.id.polygon = "polygon.id",
  b.show.polygon.label = TRUE,
  v.info.tag.polygon.label = NULL,
  v.info.veld.polygon.label = NULL,
  b.show.polygon.popup = TRUE,
  v.info.tag.polygon.popup = NULL,
  v.info.veld.polygon.popup = NULL,
  c.fill.numeric.polygon = NULL,
  n.fill.opacity.polygon = 0.3,
  c.stroke.color.polygon = "black",
  n.stroke.weight.polygon = 1,
  b.show.stroke.polygon = TRUE,
  df.text = NULL,
  v.coord.text = c("text.lon", "text.lat"),
  c.text.label = "text.label"
)
}
\arguments{
\item{c.leaflet.title}{Leaflet title; also used for filename.}

\item{c.legend.title}{Legend title.}

\item{c.layer}{What feature should be used for layering? (default: NULL).}

\item{v.coord.point.midpoint}{Option to set the midpoint, else the center of the data will chosen (default: NULL).}

\item{n.zoom}{At what zoom level should leaflet be opened? (default: 13).}

\item{b.save.leaflet}{Should we save the leaflet? (default: FALSE).}

\item{b.add.streetsmart}{Should we add StreetSmart link to pop-up? (default: TRUE).}

\item{df.point}{POINT - (default: NULL).}

\item{v.coord.point}{POINT -  (default: c("bord.lon", "bord.lat")).}

\item{c.id.point}{POINT - (default: "bord.id").}

\item{b.show.point.label}{POINT - (default: TRUE).}

\item{v.info.tag.point.label}{POINT - (default: NULL).}

\item{v.info.veld.point.label}{POINT -  (default: NULL).}

\item{b.show.point.popup}{POINT - (default: TRUE).}

\item{v.info.tag.point.popup}{POINT - (default: NULL).}

\item{v.info.veld.point.popup}{POINT - (default: NULL).}

\item{df.fill.point}{POINT -  (default: NULL).}

\item{c.fill.factor.point}{POINT -  (default: NULL).}

\item{c.fill.numeric.point}{POINT - (default: NULL).}

\item{df.weight.point}{POINT -  (default: NULL).}

\item{c.weight.point}{POINT -  (default: NULL).}

\item{df.stroke.color.point}{POINT -   (default: NULL).}

\item{c.stroke.factor.point}{POINT -    (default: NULL).}

\item{c.stroke.numeric.point}{POINT -   (default: NULL}

\item{df.stroke.weight.point}{POINT -   (default: NULL).}

\item{c.stroke.weight.point}{POINT -    (default: NULL).}

\item{n.opacity.fill}{POINT - opacity of fill (default: 0.9).}

\item{n.opacity.stroke}{POINT - Opacity of stroke (default: 0.9).}

\item{df.line}{LINE - (default: NULL).}

\item{v.coord.line}{LINE - (default: c("line.lon", "line.lat")).}

\item{c.id.line}{LINE - (default: "line.id").}

\item{b.show.line.label}{LINE - (default: FALSE).}

\item{v.info.tag.line.label}{LINE - (default: NULL).}

\item{v.info.veld.line.label}{LINE - (default: NULL).}

\item{b.show.line.popup}{LINE - (default: FALSE).}

\item{v.info.tag.line.popup}{LINE (default: NULL).}

\item{v.info.veld.line.popup}{LINE - (default: NULL).}

\item{df.color.line}{LINE - (default: NULL).}

\item{c.color.line}{LINE - (default: NULL).}

\item{df.weight.line}{LINE - (default: NULL).}

\item{c.weight.line}{LINE - (default: NULL).}

\item{n.opacity.line}{LINE -  (default: 1).}

\item{df.polygon}{POLYGON - (default: NULL).}

\item{v.coord.polygon}{POLYGON - (default: c("polygon.lon", "polygon.lat")).}

\item{c.id.polygon}{POLYGON - (default: "polygon.id").}

\item{b.show.polygon.label}{POLYGON - (default: TRUE).}

\item{v.info.tag.polygon.label}{POLYGON - (default: NULL).}

\item{v.info.veld.polygon.label}{POLYGON - (default: NULL).}

\item{b.show.polygon.popup}{POLYGON - (default: TRUE).}

\item{v.info.tag.polygon.popup}{POLYGON - (default: NULL).}

\item{v.info.veld.polygon.popup}{POLYGON - (default: NULL).}

\item{c.fill.numeric.polygon}{POLYGON -  (default: NULL).}

\item{n.fill.opacity.polygon}{POLYGON - (default: 0.3).}

\item{c.stroke.color.polygon}{POLYGON - (default: "black").}

\item{n.stroke.weight.polygon}{POLYGON - (default: 1).}

\item{b.show.stroke.polygon}{POLYGON - (default: TRUE).}

\item{df.text}{TEXT - (default: NULL).}

\item{v.coord.text}{TEXT - (default: c("text.lon", "text.lat")).}

\item{c.text.label}{TEXT - (default: "text.label").}

\item{v.add.date, v.add.time}{Vector of booleans to specify whether date and/or time should be added.
This should be as long as x (default: TRUE and TRUE, resp.).}
}
\value{
A beautiful leaflet!
}
\description{
Plots leaflet.
}
\details{
-
}
\examples{
f_plot_leaflet(

     c.leaflet.title           = "Great Leaflet",
     c.legend.title            = "Legend:",
     c.layer                   = NULL,
     v.coord.point.midpoint    = NULL,
     n.zoom                    = 13,
     b.save.leaflet            = FALSE,
     b.add.date                = TRUE,
     b.add.time                = TRUE,
     b.add.streetsmart         = TRUE,
     df.point                  = NULL,
     v.coord.point             = c("bord.lon", "bord.lat"),
     c.id.point                = "bord.id",
     b.show.point.label        = TRUE,
     v.info.tag.point.label    = NULL,
     v.info.veld.point.label   = NULL,
     b.show.point.popup        = TRUE,
     v.info.tag.point.popup    = NULL,
     v.info.veld.point.popup   = NULL,
     df.fill.point             = NULL,
     c.fill.factor.point       = NULL,
     c.fill.numeric.point      = NULL,
     df.weight.point           = NULL,
     c.weight.point            = NULL,
     df.stroke.color.point     = NULL,
     c.stroke.factor.point     = NULL,
     c.stroke.numeric.point    = NULL,
     df.stroke.weight.point    = NULL,
     c.stroke.weight.point     = NULL,
     n.opacity.fill            = 0.9,
     n.opacity.stroke          = 0.9,
     df.line                   = NULL,
     v.coord.line              = c("line.lon", "line.lat"),
     c.id.line                 = "line.id",
     b.show.line.label         = FALSE,
     v.info.tag.line.label     = NULL,
     v.info.veld.line.label    = NULL,
     b.show.line.popup         = FALSE,
     v.info.tag.line.popup     = NULL,
     v.info.veld.line.popup    = NULL,
     df.color.line             = NULL,
     c.color.line              = NULL,
     df.weight.line            = NULL,
     c.weight.line             = NULL,
     n.opacity.line            = 1,
     df.polygon                = NULL,
     v.coord.polygon           = c("polygon.lon", "polygon.lat"),
     c.id.polygon              = "polygon.id",
     b.show.polygon.label      = TRUE,
     v.info.tag.polygon.label  = NULL,
     v.info.veld.polygon.label = NULL,
     b.show.polygon.popup      = TRUE,
     v.info.tag.polygon.popup  = NULL,
     v.info.veld.polygon.popup = NULL,
     c.fill.numeric.polygon    = NULL,
     n.fill.opacity.polygon    = 0.3,
     c.stroke.color.polygon    = "black",
     n.stroke.weight.polygon   = 1,
     b.show.stroke.polygon     = TRUE,
     df.text                   = NULL,
     v.coord.text              = c("text.lon", "text.lat"),
     c.text.label              = "text.label"
)
}
\author{
Pieter Overdevest
}
