% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Reactable.R
\name{f_reactable}
\alias{f_reactable}
\title{Create reactable table in html document}
\usage{
f_reactable(
  df.input,
  c.col.default.align = "center",
  n.col.default.max.width = 500,
  v.col.text = NULL,
  v.col.text.name = NULL,
  v.col.text.align = NULL,
  v.col.text.width = NULL,
  v.col.digit = NULL,
  v.col.digit.name = NULL,
  v.col.digit.number = NULL,
  v.col.digit.align = NULL,
  v.col.digit.width = NULL,
  v.col.euro = NULL,
  v.col.euro.name = NULL,
  v.col.euro.number = NULL,
  v.col.euro.align = NULL,
  v.col.euro.width = NULL,
  v.col.link.label = NULL,
  v.col.link.url = NULL,
  v.col.link.name = NULL,
  v.col.link.align = NULL,
  v.col.link.width = NULL,
  v.row.number.color = NULL,
  c.row.number.color = NULL,
  n.defaultPageSize = 10,
  b.showPageSizeOptions = TRUE,
  v.pageSizeOptions = c(10, 20, 30),
  b.filterable = TRUE,
  b.searchable = TRUE
)
}
\arguments{
\item{df.input}{Data frame to print as rectable table in html document.}

\item{c.col.default.align}{What is default alignment. Options are: "left", "center", and "right" (default: "center").}

\item{n.col.default.max.width}{What is the max width of a column (default: 500).}

\item{v.col.text}{Vector with feature names that should be formated as text (default: NULL).}

\item{v.col.text.name}{Vector with header names that should be used in the table instead (default: NULL).}

\item{v.col.text.align}{Vector with alignment of the text. Options are: "left", "center", and "right" (default: NULL).}

\item{v.col.text.width}{Vector with widths of the concerned columns (default: NULL).}

\item{v.col.digit}{Vector with feature names that should be formated as number (default: NULL).}

\item{v.col.digit.name}{Vector with header names that should be used in the table instead (default: NULL).}

\item{v.col.digit.number}{Vector with number of digits to use (default: NULL).}

\item{v.col.digit.align}{Vector with alignment of the numbers. Options are: "left", "center", and "right" (default: NULL).}

\item{v.col.digit.width}{Vector with widths of the concerned columns (default: NULL).}

\item{v.col.euro}{Vector with feature names that should be formated as euro (default: NULL).}

\item{v.col.euro.name}{Vector with header names that should be used in the table instead (default: NULL).}

\item{v.col.euro.number}{Vector with number of digits to use (default: NULL).}

\item{v.col.euro.align}{Vector with alignment of the euro's. Options are: "left", "center", and "right" (default: NULL).}

\item{v.col.euro.width}{Vector with widths of the concerned columns (default: NULL).}

\item{v.col.link.label}{Vector with feature names that should be used as link label (default: NULL).}

\item{v.col.link.url}{Vector with feature names that should be used as link url (default: NULL).}

\item{v.col.link.name}{Vector with header names that should be used in the table instead (default: NULL).}

\item{v.col.link.align}{Vector with alignment of the links. Options are: "left", "center", and "right" (default: NULL).}

\item{v.col.link.width}{Vector with widths of the concerned columns (default: NULL).}

\item{v.row.number.color}{Row numbers that need to be colored with c.row.number.color (default: NULL).}

\item{c.row.number.color}{Color to give to the concerned rows (default: NULL).}

\item{n.defaultPageSize}{What is the default page size? (default: 10).}

\item{b.showPageSizeOptions}{Should we show page size options? (default: TRUE).}

\item{v.pageSizeOptions}{What page size options to provide (default: c(10, 20, 30)).}

\item{b.filterable}{Should table be filterable? (default: TRUE).}

\item{b.searchable}{Should table be searchable? (default: TRUE).}
}
\value{
Reactable table.
}
\description{
Creates reactable table in html document.
}
\details{
-
}
\examples{
f_reactable(

     df.input                = mtcars,
     c.col.default.align     = "center",
     n.col.default.max.width = 500,
     v.col.text              = NULL,
     v.col.text.name         = NULL,
     v.col.text.align        = NULL,
     v.col.text.width        = NULL,
     v.col.digit             = NULL,
     v.col.digit.name        = NULL,
     v.col.digit.number      = NULL,
     v.col.digit.align       = NULL,
     v.col.digit.width       = NULL,
     v.col.euro              = NULL,
     v.col.euro.name         = NULL,
     v.col.euro.number       = NULL,
     v.col.euro.align        = NULL,
     v.col.euro.width        = NULL,
     v.col.link.label        = NULL,
     v.col.link.url          = NULL,
     v.col.link.name         = NULL,
     v.col.link.align        = NULL,
     v.col.link.width        = NULL,
     v.row.number.color      = NULL,
     c.row.number.color      = NULL,
     n.defaultPageSize       = 10,
     b.showPageSizeOptions   = TRUE,
     v.pageSizeOptions       = c(10, 20, 30),
     b.filterable            = TRUE,
     b.searchable            = TRUE
)
}
\author{
Pieter Overdevest
}
