% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Read data from file.R
\name{f_read_data_from_file}
\alias{f_read_data_from_file}
\title{Read data from file}
\usage{
f_read_data_from_file(
  v.file.string = NULL,
  c.file.type,
  c.path = NULL,
  b.exact.match = FALSE,
  c.file.string.exclude = NULL,
  b.show.info = FALSE,
  c.sheet.name = NULL,
  n.skip.rows = 0,
  c.delim = NULL,
  c.decimal.mark = ".",
  b.col.names = TRUE,
  l.col.type = NULL,
  n.guess.max = 1000,
  c.table.name = NULL,
  c.show.report = "all",
  b.add.mod.date.path.file = FALSE,
  b.clean.up.header.names = TRUE
)
}
\arguments{
\item{v.file.string}{String to search for in the filename. In case of Google Sheet this can be kept NULL (default: NULL).}

\item{c.file.type}{Options: 'xls', 'csv', 'tsv', 'txt', 'delim', 'rds', 'fst', 'xml', 'parquet', 'sqlite', 'shp', 'dbf', 'gs'}

\item{c.path}{Path where file should be searched for (default: NULL).}

\item{b.exact.match}{Exact match or not (default: FALSE).}

\item{c.file.string.exclude}{Means to filter out files by providing a string that should not exist in the filename
(default: NULL).}

\item{b.show.info}{Should we show info? (default: FALSE).}

\item{c.sheet.name}{In case of xls (default: NULL).}

\item{n.skip.rows}{In case of xls (default: 0).}

\item{c.delim}{In case of txt, csv, delim (default: NULL).}

\item{c.decimal.mark}{In case of txt, csv, delim (default: '.').}

\item{b.col.names}{Do the data contain header names? (default: TRUE).}

\item{l.col.type}{List with column types, see also Details (default: NULL).}

\item{n.guess.max}{Number of rows to check on data type (default: 1000).}

\item{c.table.name}{In case of SQLITE (default: NULL).}

\item{c.show.report}{Options: "none", "minimal", "all" (default: "all").}

\item{b.add.mod.date.path.file}{Should we add auxiliry features: mod.date, path.file? (default: FALSE).}

\item{b.clean.up.header.names}{Should we clean up the header names? (default: TRUE).}
}
\value{
Data frame containing the data read from the file.
}
\description{
Reads data from file.
}
\details{
In case of Excel file, l.col.type should consist of list of: "skip", "guess", "logical", "numeric", "date",
"text" or "list". In case of delimited file, l.col.type should consist of list of: c = character, i = integer,
n = number, d = double, l = logical, D = date, T = date time, t = time, ? = guess, or _/- to skip the column.
}
\examples{
f_read_data_from_file(

     v.file.string            = "bord.allocation",
     c.file.type              = "xls",
     c.path                   = path.data,
     b.exact.match            = FALSE,
     c.file.string.exclude    = NULL,
     b.show.info              = FALSE,
     c.sheet.name             = NULL,
     n.skip.rows              = 0,
     c.delim                  = NULL,
     c.decimal.mark           = ".",
     b.col.names              = TRUE,
     l.col.type               = NULL,
     n.guess.max              = 1000,
     c.table.name             = NULL,
     c.show.report            = "all",
     b.add.mod.date.path.file = FALSE,
     b.clean.up.header.names  = TRUE
)
}
\author{
Pieter Overdevest
}
