% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Render Quarto.R
\name{f_render_quarto}
\alias{f_render_quarto}
\title{Render Quarto document}
\usage{
f_render_quarto(
  c.folder.source,
  c.file.source,
  c.folder.destination,
  c.file.destination = NULL,
  b.add.date = TRUE,
  b.add.time = FALSE
)
}
\arguments{
\item{c.folder.source}{Folder holding the qmd file.}

\item{c.file.source}{Core name of the qmd file.}

\item{c.folder.destination}{Folder where to write the rendered file.}

\item{c.file.destination}{Option to give the rendered file a new name. By default c.file.source is used (default: NULL).}

\item{b.add.date, b.add.time}{Should we add date and/or time to the filename (default: TRUE and FALSE, resp.).}
}
\value{
It saves the rendered file; nothing is returned from this function directly.
}
\description{
Renders Quarto document.
}
\details{
-
}
\examples{
f_render_quarto(

     c.folder.source      = path.code,
     c.file.source        = "Q - Data Understanding Clickstream - HTML.qmd",
     c.folder.destination = path.deliverables,
     c.file.destination   = NULL,
     b.add.date           = TRUE,
     b.add.time           = FALSE
)
}
\author{
Pieter Overdevest
}
