% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Search values in list of data frames.R
\name{f_search_values_in_list_of_data_frames}
\alias{f_search_values_in_list_of_data_frames}
\title{Search values in list of data frames}
\usage{
f_search_values_in_list_of_data_frames(
  l.input,
  v.element,
  l.search,
  v.strategy = NULL
)
}
\arguments{
\item{l.input}{List of data frames, in which we want to determine which data frame(s) contain the requested values.}

\item{v.element}{Vector of volumn name(s) in the data frames in which we search for requested values.}

\item{l.search}{List of vectors with value(s) to search for in said column(s).}

\item{v.strategy}{Vector of "all" and/or "any", corresponding to each column name that is searched in. This allows to
match all values or to match any of the values in the respective vector in l.search. In case of the default (NULL)
all values in the respective vector in l.search must match.}
}
\value{
Vector of list indices that contain requested values.
}
\description{
Search values in list of data frames.
}
\details{
Note, v.element, l.search, and v.strategy (if given) must have equal length.
}
\examples{
v.output <- f_search_values_in_list_of_data_frames(

    l.input    = l.bord.allocation,
    v.element  = c("bord.type", "jaar.laatste.vastlegging"),
    l.search   = list(c("A01120", "A01100"), "2021"),
    v.strategy = NULL
)
}
\author{
Pieter Overdevest
}
