% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Source.R
\name{f_source}
\alias{f_source}
\title{Source latest file.}
\usage{
f_source(script.to.source, path)
}
\arguments{
\item{script.to.source}{String in filename to source.}

\item{path}{Full path to folder containing the file.}
}
\value{
None
}
\description{
Sources latest file.
}
\details{
-
}
\examples{
f_source(
       script.to.source = "I - IWD",
       path             = paste0(path.root.local, "IWD/Projects/Initialization IWD")
)
}
\author{
Pieter Overdevest
}
