% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Summarize.R
\name{f_summarize}
\alias{f_summarize}
\title{Give basic info on columns in data frame}
\usage{
f_summarize(
  df.input,
  b.sort = TRUE,
  b.view = TRUE,
  b.return = FALSE,
  n.char = "all",
  n.freq = 3
)
}
\arguments{
\item{df.input}{Data frame to summarize.}

\item{b.sort}{Should items be sorted? (default: TRUE).}

\item{b.view}{Should we display the result? (default: TRUE).}

\item{b.return}{Should we return the resulting data frame (default: FALSE).}

\item{n.char}{Number of characters to show of each item (default: "all").}

\item{n.freq}{Number of items to show (default: 3).}
}
\value{
Data frame summarizing the data in df.input, in case b.return is TRUE.
}
\description{
Gives basic info on columns in data frame.
}
\details{
-
}
\examples{
f_summarize(

     df.input = mtcars,
     b.sort   = TRUE,
     b.view   = TRUE,
     b.return = FALSE,
     n.char   = "all",
     n.freq   = 3
)
}
\author{
Pieter Overdevest
}
