% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Table.R
\name{f_table}
\alias{f_table}
\title{Prepare 2D frequency table}
\usage{
f_table(
  df.input,
  c.ver,
  c.hor = NULL,
  c.useNA = "ifany",
  c.type = "abs",
  n.round = 1,
  b.warning = TRUE
)
}
\arguments{
\item{df.input}{Data frame containing features c.ver (and c.hor)}

\item{c.ver}{Feature name to pivot vertically.}

\item{c.hor}{Feature name to pivot horizontally. If provided this will result in a 2D frequency table, and when not
provided the result will be a 1D frequency table (default: NULL).}

\item{c.useNA}{How to deal with NAs. Options are: "ifany", "no" "always" (default: "ifany").}

\item{c.type}{Type result. Options are: "abs", "rel" (default: "abs").}

\item{n.round}{Rounding the numbers (default: 1).}

\item{b.warning}{Should warnings be shown? (default: TRUE).}
}
\value{
A 2D (or 1D) frequency table.
}
\description{
Prepares 2D frequency table.
}
\details{
-
}
\examples{
f_table(

      df.input  = mtcars,
      c.ver     = "gear",
      c.hor     = "am",
      c.useNA   = "ifany",
      c.type    = "abs",
      n.round   = 1,
      b.warning = TRUE
)
}
\author{
Pieter Overdevest
}
