% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Unique.R
\name{f_unique}
\alias{f_unique}
\title{Return sorted unique elements}
\usage{
f_unique(
  v.vector,
  b.show.freq = FALSE,
  b.sort.by.val = FALSE,
  b.sort.by.freq = FALSE,
  n.char = "all"
)
}
\arguments{
\item{v.vector}{Vector with elements to get unique values of.}

\item{b.show.freq}{Should we add the frequency of values occuring? (default: FALSE).}

\item{b.sort.by.val}{Should results be sorted by values? (default: FALSE).}

\item{b.sort.by.freq}{Should results be sorted by frequency? (default: FALSE).}

\item{n.char}{The number of characters to use of each value (default: "all").}
}
\value{
Concatenated string of items.
}
\description{
Returns sorted unique elements.
}
\details{
-
}
\examples{
v.result <- f_unique(

     v.vector       = c(50, 30, 100, 100, 100, 30),
     b.show.freq    = FALSE,
     b.sort.by.val  = FALSE,
     b.sort.by.freq = FALSE,
     n.char         = "all"
)
}
\author{
Pieter Overdevest
}
