% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Upload file to Google Drive.R
\name{f_upload_file_to_google_drive}
\alias{f_upload_file_to_google_drive}
\title{Upload given file to Google Drive}
\usage{
f_upload_file_to_google_drive(
  c.file.string,
  c.file.type,
  c.path,
  c.folder,
  b.exact.match = FALSE,
  c.file.string.exclude = NULL,
  c.show.report = "all"
)
}
\arguments{
\item{c.file.string}{Core name of the file to upload.}

\item{c.file.type}{File type of file to upload.}

\item{c.path}{Path holding the file to upload.}

\item{c.folder}{Google folder where to upload the file.}

\item{b.exact.match}{Should file name have exact match with c.file.string (default: FALSE)}

\item{c.file.string.exclude}{Any string to exclude files for (default: NULL).}

\item{c.show.report}{What info should be printed to console when searching for files (default: "all")}
}
\value{
Nothing.
}
\description{
Uploads given file to Google Drive.
}
\details{
-
}
\examples{
f_upload_file_to_google_drive(

     c.file.string         = "Daisycon Feed",
     c.file.type           = "csv",
     c.path                = path.deliverables,
     c.folder              = "12. Data Science BUHV-Oletti/",
     b.exact.match         = FALSE,
     c.file.string.exclude = NULL,
     c.show.report         = "all"
)
}
\author{
Pieter Overdevest
}
