% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Vector info.R
\name{f_vector_info}
\alias{f_vector_info}
\title{Provide information on vector}
\usage{
f_vector_info(v.input, name, n.top, show.freq, n.width, c.sort.by)
}
\arguments{
\item{v.input}{Vector of items to check statistics of.}

\item{name}{Name to print above the table.}

\item{n.top}{Max number of items to show in the list.}

\item{show.freq}{Should frequency be shown?}

\item{n.width}{Number of characters of the items to show in the list.}

\item{c.sort.by}{How to sort the items in the frequency table, by its 'frequency' or 'value' (default: 'frequency').}
}
\value{
Nothing. Only prints to console.
}
\description{
Provides informartion on vector. Typically, you will use f_info() and not use f_vector_info().
The function f_info() calls f_vector_info() for each feature in the object entered in f_info().
}
\details{
-
}
\author{
Pieter Overdevest
}
