% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Welcome.R
\name{f_welcome}
\alias{f_welcome}
\title{Welcome!}
\usage{
f_welcome(x = "Pieter")
}
\arguments{
\item{x}{Name to welcome (default: "Pieter").}
}
\value{
String of welcome!
}
\description{
Gives welcome.
}
\details{
-
}
\examples{
f_welcome(

    x = "Pieter"
)
}
\author{
Pieter Overdevest
}
