% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Write data to file merged.R
\name{f_write_data_to_file_merged}
\alias{f_write_data_to_file_merged}
\title{Reads files, appends data and writes merged file}
\usage{
f_write_data_to_file_merged(
  df.input,
  c.path,
  c.file.string,
  c.file.type,
  c.sheet.name = "Sheet1",
  v.add.date = TRUE,
  v.add.time = FALSE,
  n.day.max = 1e+06
)
}
\arguments{
\item{df.input}{Data frame with data to append to merged data,}

\item{c.path}{Path where file(s) are searched for.}

\item{c.file.string}{Core of the filename of the files to be merged. The same is used for the file that is written.}

\item{c.file.type}{File type of files to merge. The merged file will also be saved in this format.}

\item{c.sheet.name}{Sheetname containing the data to be merged. Is relevant in case of c.file.type being xls
(default: "Sheet1").}

\item{v.add.date, v.add.time}{Vector of booleans to specify whether data and/or time should be added.
This should be as long as x (default: TRUE and FALSE, resp.).}

\item{n.day.max}{Files newer than the number of days given in n.day.max are merged. The high default value essentially
means that all files are merged (default: 1000000).}
}
\value{
Boolean whether the vector contains unique values.
}
\description{
Reads, appends and writes data to merged file. Function looks for files with same core name in folder -
usually only the date (and time) differs. The files can be filtered on their age (number of days). Only files older
than n.day.max are merged and written to a new file. New data can be appended through df.input. The files that were
merged are moved to an Archive\ folder.
}
\details{
-
}
\examples{
f_write_data_to_file_merged(

     df.input       = mtcars,
     c.path         = path.data,
     c.file.string  = "dummy file",
     c.sheet.name   = "Sheet1",
     c.file.type    = "xls",
     v.add.date     = TRUE,
     v.add.time     = FALSE,
     n.day.max      = 1000000
)
}
\author{
Pieter Overdevest
}
