% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generaltoolbox.R
\docType{package}
\name{generaltoolbox-package}
\alias{generaltoolbox}
\alias{generaltoolbox-package}
\title{generaltoolbox: General toolbox by Pieter Overdevest}
\description{
This package contains general toolbox functions.
}
\keyword{internal}
