% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Send slack comment.R
\name{f_send_slack_comment}
\alias{f_send_slack_comment}
\title{Send Slack Comment.}
\usage{
f_send_slack_comment(c.slack.hook, c.message)
}
\arguments{
\item{c.slack.hook}{String with slack hook.}

\item{c.message}{Message to communicate.}
}
\value{
None
}
\description{
Sends Slack comment.
}
\details{
How to obtain a Slack Hook:

* Go to https://api.slack.com/apps

* If needed, create a new app

* Go to ‘Incoming Webhooks’

* If not already, activate Incoming Webhooks

* Add new webhook to workspace

* Select channel

* Copy webhook and use for POST() function.

* Add app to Slack by ‘Add apps’, or the channel may already be pointing to it for you
}
\examples{
f_send_slack_comment(
       c.slack.hook = "https://hooks.slack.com/services/........./......../...........",
       c.message    = "Hello World!"
)
}
\author{
Pieter Overdevest
}
