% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f - Vector info.R
\name{f_vector_info}
\alias{f_vector_info}
\title{Provide information on vector}
\usage{
f_vector_info(v.input, name, n.top, show.freq, c.sort.by, n.width)
}
\arguments{
\item{v.input}{Vector of items to check statistics of.}

\item{name}{Name to print above the table.}

\item{n.top}{Max number of items to show in the list.}

\item{show.freq}{Should frequency be shown?}

\item{c.sort.by}{How to sort the items in the frequency table, by its 'frequency' or 'value' (default: 'frequency').}

\item{n.width}{Number of characters of the items to show in the list.}
}
\value{
Nothing. Only prints to console.
}
\description{
Provides informartion on vector. Typically, you will use f_info() and not use f_vector_info().
The function f_info() calls f_vector_info() for each feature in the object entered in f_info().
}
\details{
-
}
\examples{
f_vector_info(

     v.input   = c(0, 2, 2, NA, NA, NA, 0/0, -0/0, -0/0, 0/0, 6/0, -7/0, -8/0, 9/0, 10/0),
     name      = "Pieter",
     n.top     = 10,
     show.freq = TRUE,
     c.sort.by = 'frequency',
     n.width   = 29
)
}
\author{
Pieter Overdevest
}
